+++ NEWS.md
@@ -17,6 +17,7 @@
 
 ### Miscellaneous
 
+* New Session command (create new R session with same project or working directory)
 * Raise limit on shinyapps uploads to 1GB from 100MB
 * 'Edit -> Replace and Find' opens Find toolbar if not already open (e.g. with Cmd+Shift+J)
 * Improve performance of console for large and/or rapdily updating output